+++ lib/liquid/tags/include.rb
@@ -42,14 +42,15 @@ module Liquid
 
     def render(context)
       template_name = context.evaluate(@template_name_expr)
-      partial = load_cached_partial(template_name, context)
+      raise ArgumentError.new(options[:locale].t("errors.argument.include")) unless template_name
 
+      partial = load_cached_partial(template_name, context)
       context_variable_name = template_name.split('/'.freeze).last
 
       variable = if @variable_name_expr
         context.evaluate(@variable_name_expr)
       else
-        context.find_variable(template_name)
+        context.find_variable(template_name, raise_on_not_found: false)
       end
 
       old_template_name = context.template_name